requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.6
flake8==3.9.2
idna==3.2
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
PyJWT==1.7.1
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
twilio==6.63.2
urllib3==1.26.6
